refactor: rename chat components to lowercase
