+++ src/App.js
@@ -33,7 +33,7 @@ import JoinTribe from "./components/join-tribe/join-tribe-page"
 import TribeProfile from "./components/tribes/tribe-profile/tribe-profile"
 import "./index.scss";
 import Footer from './components/footer/footer';
-import Chat from './components/chat/Chat';
+import Chat from './components/chat/chat';
 
 const ProtectedRoute = ({ component: Component, path, ...rest }) => {
   const { pathname, search } = useLocation();
