+++ src/components/chat/chat.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import RoomList from './room-list/RoomList';
-import ChatRoom from './chat-room/ChatRoom';
+import RoomList from './room-list/room-list';
+import ChatRoom from './chat-room/chat-room';
 
 const Chat = () => {
   const { isVisibleChat } = useSelector(state => state.chat);
